+++ yash-env/CHANGELOG.md
@@ -17,6 +17,7 @@ A _private dependency_ is used internally and not visible to downstream users.
   workaround for the now non-functional
   `<system::SharedSystem as System>::new_child_process`.
 - The following traits have been added to the `system` module:
+    - `Close`: Declares the `close` method for closing file descriptors.
     - `Dup`: Declares the `dup` and `dup2` methods for duplicating file
       descriptors.
     - `Fstat`: Declares `fstat` and `fstatat` methods for getting file